backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
google-genai>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
